requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
boto3==1.34.34
python-dotenv==1.0.0
httpx==0.26.0
slowapi==0.1.9

